% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calib}
\alias{calib}
\alias{valid}
\title{Rainfall and streamflow data}
\format{data.frame with numerical vectors for Q and P}
\source{
Some hydrology course at Potsdam University
}
\description{
Timeseries of catchment precipitation  (P, mm/h) and discharge, which is streamflow at the outlet (Q, m^3/s).
The data is split up in a calibration and validation period and consists of
two data.frames: \code{calib} and \code{valid}.\cr
The catchment is small, approcimately 1.6 km^2.
}
\examples{

\donttest{
data(calib)
data(valid)
head(calib)
}

}
\seealso{
\code{\link{lsc}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, 2013
}
\keyword{datasets}
