% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsc.R
\name{lsc}
\alias{lsc}
\title{Linear storage cascade, unit hydrograph}
\usage{
lsc(P, Q, area = 50, Qbase = Q[1], n = 2, k = 3, x = 1:length(P),
  fit = 1:length(Q), plot = TRUE, main = "Precipitation and discharge",
  plotsim = TRUE, returnsim = FALSE, type = c("o", "l"),
  legx = "center", legy = NULL, ...)
}
\arguments{
\item{P}{Vector with precipitation values \bold{in mm in hourly spacing}}

\item{Q}{Vector with observed discharge (runoff) \bold{in m^3/s}
with the same length as precipitation.}

\item{area}{Single numeric. Catchment area \bold{in km^2}}

\item{Qbase}{Baseflow that is added to UH-induced simulated Q
DEFAULT: Q[1] thus cutting off baseflow in a very simple manner.}

\item{n}{Numeric. Initial number of storages in cascade.
Not necessarily integer. DEFAULT: 2}

\item{k}{Numeric. Initial storage coefficient (resistance to let water run out).
High damping, slowly reacting landscape, high k. DEFAULT: 3}

\item{x}{Vector for the x-axis of the plot. DEFAULT: sequence along P}

\item{fit}{Integer vector. Indices for a subset of Q that Qsim is fitted to.
DEFAULT: all of Q}

\item{plot}{Logical. plot input data? DEFAULT: TRUE}

\item{main}{Character string. DEFAULT: "Precipitation and discharge"}

\item{plotsim}{Logical. add best fit to plot? DEFAULT: TRUE}

\item{returnsim}{Logical. Return simulated Q instead of parameters of UH?
DEFAULT: FALSE}

\item{type}{Vector with two characters: type as in \code{\link{plot}},
repeated if only one is given. 1st for obs, 2nd for sim.
DEFAULT: c("o","l")}

\item{legx}{legend position. DEFAULT: "center"}

\item{legy}{legend position. DEFAULT: NULL}

\item{\dots}{Arguments passed to optim}
}
\value{
\emph{Either} vector with optimized n and k and the Nash-Sutcliffe Index,
        \emph{or} simulated discharge, depending on the value of \code{returnsim}
}
\description{
Optimize the parameters for unit hydrograph as in the framework of the
linear storage cascade. Plot observed & simulated data
}
\examples{
# Simplified for this course. True documentation is in berryFunctions::lsc
\donttest{ # this will not run directly, as the debugging exercise requires to fix the bugs
data(calib)
lsc(calib$P, calib$Q, area=1.6, quietNA=T)
}

}
\references{
\url{http://ponce.sdsu.edu/onlineuhcascade.php}\cr
   Skript 'Abflusskonzentration' zur Vorlesungsreihe Abwasserentsorgung I von Prof. Krebs an der TU Dresden\cr
  \url{http://tu-dresden.de/die_tu_dresden/fakultaeten/fakultaet_forst_geo_und_hydrowissenschaften/fachrichtung_wasserwesen/isiw/sww/lehre/dateien/abwasserbehandlung/uebung_ws09_10/uebung_awe_1_abflusskonzentration.pdf}\cr
  \url{http://www.uni-potsdam.de/fs-g3/file.php?fileserver=klausuren&file=\%2FMaster_of_Science\%2FHydroII_Lernzettel.pdf}
}
\seealso{
\code{\link{unitHydrograph}}, \code{\link{superPos}}, \code{\link{nse}}, \code{\link{rmse}}.
         \code{deconvolution.uh} in the package hydromad, \url{http://hydromad.catchment.org}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, July 2013
}
\keyword{hplot}
\keyword{optimize}
\keyword{ts}
