% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{nse}
\alias{nse}
\title{Nash Sutcliffe efficiency}
\usage{
nse(obs, sim)
}
\arguments{
\item{obs}{vector with observed values. Should be numerical.}

\item{sim}{vector with observed values}
}
\value{
Single numerical value with NSE value
}
\description{
Nash Sutcliffe efficiency
}
\examples{
x <- c(5,7,8)
y <- c(6,7,8)
#nse(x,y)


}
