% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitHydrograph.R
\name{unitHydrograph}
\alias{unitHydrograph}
\title{Unit Hydrograph}
\usage{
unitHydrograph(n, k, t, force = FALSE)
}
\arguments{
\item{n}{Numeric. Number of storages in cascade.}

\item{k}{Numeric. Storage coefficient [1/s] (resistance to let water run out). High damping = slowly reacting landscape = high soil water absorbtion = high k.}

\item{t}{Numeric, possibly a vector. Time [s].}

\item{force}{Logical: Force the integral of the hydrograph to be 1? DEFAULT: FALSE}
}
\value{
Vector with the unit hydrograph along t
}
\description{
Calculate continuous unit hydrograph with given n and k (in the framework of the linear storage cascade)
}
\note{
The sum under the UH should always be 1 (if t is long enough). This needs yet to be checked...
}
\examples{

Time <- 0:100
plot(Time, unitHydrograph(n=2,  k=3, t=Time), type="l", las=1,
     main="Unit Hydrograph - linear storage cascade")
lines(Time, unitHydrograph(n=2,  k=8, t=Time), col=2)
lines(Time, unitHydrograph(n=5.5,k=8, t=Time), col=4)
text(c(12, 20, 50), c(0.1, 0.04, 0.025), c("n=2, k=3","n=2, k=8","n=5.5, k=8"),
     col=c(1,2,4), adj=0)

}
\seealso{
\code{\link{lsc}} on how to estimate n and k for a given discharge dataset. \code{deconvolution.uh} in the package hydromad, \url{http://hydromad.catchment.org}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, July 2013
}
\keyword{hplot}
\keyword{ts}
