% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error}
\usage{
rmse(a, b, quiet = FALSE)
}
\arguments{
\item{a}{observed}

\item{b}{predicted}

\item{quiet}{logical variable indicating whether warning messages are printed}
}
\value{
Single numerical value with NSE value
}
\description{
Root Mean Square Error
}
